refactor: cleanup some code
