+++ app/src/main/java/com/ecosense/android/featDiseaseRecognition/domain/model/RecognisedDisease.kt
@@ -1,4 +1,4 @@
-package com.ecosense.android.featDiseaseRecognition.domain
+package com.ecosense.android.featDiseaseRecognition.domain.model
 
 data class RecognisedDisease(val label: String, val confidence: Float) {
     private val probabilityString = String.format("%.1f%%", confidence * 100.0f)
