+++ app/src/main/java/com/ecosense/android/featDiseaseRecognition/presentation/analyzer/DiseaseAnalyzer.kt
@@ -1,4 +1,4 @@
-package com.ecosense.android.featDiseaseRecognition.domain
+package com.ecosense.android.featDiseaseRecognition.presentation.analyzer
 
 import android.content.Context
 import android.graphics.Bitmap
@@ -6,19 +6,21 @@ import android.graphics.Matrix
 import androidx.camera.core.ExperimentalGetImage
 import androidx.camera.core.ImageAnalysis
 import androidx.camera.core.ImageProxy
+import com.ecosense.android.featDiseaseRecognition.presentation.util.YuvToRgbConverter
+import com.ecosense.android.featDiseaseRecognition.domain.model.RecognisedDisease
 import com.ecosense.android.ml.FlowerModel
 import org.tensorflow.lite.gpu.CompatibilityList
 import org.tensorflow.lite.support.image.TensorImage
 import org.tensorflow.lite.support.model.Model
 
 class DiseaseAnalyzer(
-    private val ctx: Context,
-    private val listener: (recognisedDisease: List<RecognisedDisease>) -> Unit
+    private val context: Context,
+    private val onAnalysisResult: (recognisedDisease: List<RecognisedDisease>) -> Unit
 ) : ImageAnalysis.Analyzer {
 
     // Initializing the flowerModel by lazy so that it runs in the same thread when the process
     // method is called.
-    private val flowerModel by lazy { FlowerModel.newInstance(ctx, options) }
+    private val flowerModel by lazy { FlowerModel.newInstance(context, options) }
 
     private val options: Model.Options = Model.Options.Builder()
         .apply {
@@ -41,13 +43,13 @@ class DiseaseAnalyzer(
             items.add(RecognisedDisease(output.label, output.score))
         }
 
-        listener(items.toList())
+        onAnalysisResult(items.toList())
 
         // Close the image,this tells CameraX to feed the next image to the analyzer
         imageProxy.close()
     }
 
-    private val yuvToRgbConverter = YuvToRgbConverter(ctx)
+    private val yuvToRgbConverter = YuvToRgbConverter(context)
     private lateinit var bitmapBuffer: Bitmap
     private lateinit var rotationMatrix: Matrix
